testlib: consolidate repository specific functions in `testlib.repo`.
* migrate functions from `utils` to `repo`.
* update corresponding test-cases.

Why ?

Decouple `tests.conftest`, which patches `cmk.utils.paths` and
`cmk.utils.versions` from `tests.testlib.utils`.

`tests.testlib.utils` makes use of `cmk.utils.versions`, which in turn
initializes details within `cmk.utils.paths`.

The aim is to patch the modules first and then use them, for consistent
unit test runs.

CMK-17561
