+++ tests/agent-integration/test_plugin_version_compatibility.py
@@ -12,7 +12,7 @@ from typing import Iterator
 import docker  # type: ignore[import-untyped]
 import pytest
 
-from tests.testlib.utils import repo_path
+from tests.testlib.repo import repo_path
 
 
 @pytest.fixture(scope="module")
