+++ tests/code_quality/file_content/test_find_debug_code_web.py
@@ -7,7 +7,7 @@ from collections.abc import Iterable
 from itertools import chain
 from pathlib import Path
 
-from tests.testlib.utils import repo_path
+from tests.testlib.repo import repo_path
 
 from ..conftest import ChangedFiles
