+++ tests/code_quality/file_content/test_find_debug_print.py
@@ -10,7 +10,7 @@ from pathlib import Path
 
 import pytest
 
-from tests.testlib.utils import repo_path
+from tests.testlib.repo import repo_path
 
 from ..conftest import ChangedFiles
