+++ tests/scripts/run-dockerized.py
@@ -25,7 +25,7 @@ from pathlib import Path
 sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.dirname(os.path.realpath(__file__)))))
 
 from tests.testlib.containers import execute_tests_in_container
-from tests.testlib.utils import current_base_branch_name
+from tests.testlib.repo import current_base_branch_name
 from tests.testlib.version import CMKVersion, version_from_env
 
 from cmk.utils.version import Edition
