+++ tests/testlib/utils.py
@@ -9,13 +9,12 @@ import os
 import re
 import shlex
 import subprocess
-import sys
 import textwrap
 
 # pylint: disable=redefined-outer-name
 import time
 from collections.abc import Callable, Iterator, Sequence
-from contextlib import contextmanager, suppress
+from contextlib import contextmanager
 from pathlib import Path
 from pprint import pformat
 from urllib.parse import urlparse
@@ -23,9 +22,11 @@ from urllib.parse import urlparse
 import pexpect  # type: ignore[import-untyped]
 import pytest
 
+from tests.testlib.repo import branch_from_env, current_branch_name, repo_path
+
 from cmk.utils.version import Edition
 
-LOGGER = logging.getLogger()
+LOGGER = logging.getLogger(__name__)
 
 
 class UtilCalledProcessError(subprocess.CalledProcessError):
@@ -54,170 +55,12 @@ class PExpectDialog:
     optional: bool = False
 
 
-def repo_path() -> Path:
-    """Returns the checkout/worktree path (in contrast to the 'git-dir')
-    same as result of `git rev-parse --show-toplevel`, but repo_path is being executed
-    quite often, so we take the not-so-portable-but-more-efficient approach.
-    """
-    return Path(__file__).resolve().parent.parent.parent
-
-
-def git_essential_directories(checkout_dir: Path) -> Iterator[str]:
-    """Yields paths to all directories needed to be accessible in order to run git operations
-    Note that if a directory is a subdirectory of checkout_dir it will be skipped"""
-
-    # path to the 'real git repository directory', i.e. the common dir when dealing with work trees
-    common_dir = (
-        (
-            checkout_dir
-            / subprocess.check_output(
-                ["git", "rev-parse", "--git-common-dir"], cwd=checkout_dir, text=True
-            ).rstrip("\n")
-        )
-        .resolve()
-        .absolute()
-    )
-
-    if not common_dir.is_relative_to(checkout_dir):
-        yield common_dir.as_posix()
-
-    # In case of reference clones we also need to access them.
-    # Not sure if 'objects/info/alternates' can contain more than one line and if we really need
-    # the parent, but at least this one is working for us
-    with suppress(FileNotFoundError):
-        with (common_dir / "objects/info/alternates").open() as alternates:
-            for alternate in (Path(line).parent for line in alternates):
-                if not alternate.is_relative_to(checkout_dir):
-                    yield alternate.as_posix()
-
-
-def git_commit_id(path: Path | str) -> str:
-    """Returns the git hash for given @path."""
-    return subprocess.check_output(
-        # use the full hash - short hashes cannot be checked out and they are not
-        # unique among machines
-        ["git", "log", "--pretty=tformat:%H", "-n1"] + [str(path)],
-        cwd=repo_path(),
-        text=True,
-    ).strip("\n")
-
-
-def qa_test_data_path() -> Path:
-    return Path(__file__).parent.parent.resolve() / Path("qa-test-data")
-
-
-def is_enterprise_repo() -> bool:
-    return (repo_path() / "omd" / "packages" / "enterprise").exists()
-
-
-def is_managed_repo() -> bool:
-    return (repo_path() / "omd" / "packages" / "managed").exists()
-
-
-def is_cloud_repo() -> bool:
-    return (repo_path() / "omd" / "packages" / "cloud").exists()
-
-
-def is_saas_repo() -> bool:
-    return (repo_path() / "omd" / "packages" / "saas").exists()
-
-
 def is_containerized() -> bool:
     return (
         os.path.exists("/.dockerenv")
         or os.path.exists("/run/.containerenv")
         or os.environ.get("CMK_CONTAINERIZED") == "TRUE"
     )
-
-
-def find_git_rm_mv_files(dirpath: Path) -> list[str]:
-    del_files = []
-
-    out = subprocess.check_output(
-        [
-            "git",
-            "-C",
-            str(dirpath),
-            "status",
-            str(dirpath),
-        ],
-        encoding="utf-8",
-    ).split("\n")
-
-    for line in out:
-        if "deleted:" in line or "renamed:" in line:
-            # Ignore files in subdirs of dirpath
-            if line.split(dirpath.name)[1].count("/") > 1:
-                continue
-
-            filename = line.split("/")[-1]
-            del_files.append(filename)
-    return del_files
-
-
-def current_branch_name() -> str:
-    branch_name = subprocess.check_output(
-        ["git", "rev-parse", "--abbrev-ref", "HEAD"], encoding="utf-8"
-    )
-    return branch_name.split("\n", 1)[0]
-
-
-def current_branch_version() -> str:
-    return subprocess.check_output(
-        [
-            "make",
-            "--no-print-directory",
-            "-f",
-            str(repo_path() / "defines.make"),
-            "print-BRANCH_VERSION",
-        ],
-        encoding="utf-8",
-    ).strip()
-
-
-def current_base_branch_name() -> str:
-    branch_name = current_branch_name()
-
-    # Detect which other branch this one was created from. We do this by going back the
-    # current branches git log one step by another and check which branches contain these
-    # commits. Only search for our main (master + major-version) branches
-    commits = subprocess.check_output(
-        ["git", "rev-list", "--max-count=30", branch_name], encoding="utf-8"
-    )
-    for commit in commits.strip().split("\n"):
-        # Asking for remote heads here, since the git repos checked out in CI
-        # do not create all branches locally
-
-        # --format=%(refname): Is not supported by all distros :(
-        #
-        # heads = subprocess.check_output(
-        #    ["git", "branch", "-r", "--format=%(refname)", "--contains", commit])
-        # if not isinstance(heads, str):
-        #    heads = heads.decode("utf-8")
-
-        # for head in heads.strip().split("\n"):
-        #    if head == "refs/remotes/origin/master":
-        #        return "master"
-
-        #    if re.match(r"^refs/remotes/origin/[0-9]+\.[0-9]+\.[0-9]+$", head):
-        #        return head
-
-        lines = subprocess.check_output(
-            ["git", "branch", "-r", "--contains", commit], encoding="utf-8"
-        )
-        for line in lines.strip().split("\n"):
-            if not line:
-                continue
-            head = line.split()[0]
-
-            if head == "origin/master":
-                return "master"
-
-            if re.match(r"^origin/[0-9]+\.[0-9]+\.[0-9]+$", head):
-                return head[7:]
-
-    LOGGER.warning("Could not determine base branch, using %s", branch_name)
-    return branch_name
 
 
 def get_cmk_download_credentials() -> tuple[str, str]:
@@ -254,14 +97,6 @@ def site_id() -> str:
     return _site_id
 
 
-def add_python_paths() -> None:
-    sys.path.insert(0, str(repo_path()))
-    if is_enterprise_repo():
-        sys.path.insert(0, os.path.join(repo_path(), "non-free", "cmc-protocols"))
-        sys.path.insert(0, os.path.join(repo_path(), "non-free", "cmk-update-agent"))
-    sys.path.insert(0, os.path.join(repo_path(), "omd/packages/omd"))
-
-
 def package_hash_path(version: str, edition: Edition) -> Path:
     return Path(f"/tmp/cmk_package_hash_{version}_{edition.long}")
 
@@ -290,14 +125,6 @@ def edition_from_env(fallback: Edition | None = None) -> Edition:
     if fallback:
         return fallback
     raise RuntimeError("EDITION environment variable, e.g. cre or enterprise, is missing")
-
-
-def branch_from_env(*, env_var: str, fallback: str | Callable[[], str] | None = None) -> str:
-    if branch := os.environ.get(env_var):
-        return branch
-    if fallback:
-        return fallback() if callable(fallback) else fallback
-    raise RuntimeError(f"{env_var} environment variable, e.g. master, is missing")
 
 
 def spawn_expect_process(
