+++ tests/unit/cmk/base/test_checks.py
@@ -3,7 +3,7 @@
 # This file is part of Checkmk (https://checkmk.com). It is subject to the terms and
 # conditions defined in the file COPYING, which is part of this source code package.
 
-from tests.testlib.utils import repo_path
+from tests.testlib.repo import repo_path
 
 
 def test_check_plugin_header() -> None:
