+++ tests/unit/cmk/gui/test_autocompleters.py
@@ -3,7 +3,7 @@
 # This file is part of Checkmk (https://checkmk.com). It is subject to the terms and
 # conditions defined in the file COPYING, which is part of this source code package.
 
-from tests.testlib.utils import is_enterprise_repo
+from tests.testlib.repo import is_enterprise_repo
 
 from cmk.gui.valuespec import autocompleter_registry
