+++ tests/unit/cmk/gui/test_i18n.py
@@ -14,7 +14,7 @@ from pathlib import Path
 import flask
 import pytest
 
-from tests.testlib.utils import repo_path
+from tests.testlib.repo import repo_path
 
 import cmk.utils.paths
