+++ tests/unit/cmk/gui/watolib/test_host_attributes.py
@@ -6,7 +6,7 @@
 # pylint: disable=redefined-outer-name
 import pytest
 
-from tests.testlib.utils import is_cloud_repo, is_enterprise_repo
+from tests.testlib.repo import is_cloud_repo, is_enterprise_repo
 
 import cmk.gui.watolib.host_attributes as attrs
