+++ tests/unit/cmk/plugins/collection/agent_based/test_if_brocade_lancom.py
@@ -630,7 +630,7 @@ if __name__ == "__main__":
     # Just run this file from your IDE and dive into the code.
     import os
 
-    from tests.testlib.utils import repo_path
+    from tests.testlib.repo import repo_path
 
     assert not pytest.main(
         [
