+++ tests/unit/cmk/plugins/proxmox_ve/agent_based/test_proxmox_ve_node_info.py
@@ -98,7 +98,7 @@ if __name__ == "__main__":
     # Just run this file from your IDE and dive into the code.
     import os
 
-    from tests.testlib.utils import repo_path
+    from tests.testlib.repo import repo_path
 
     assert not pytest.main(
         [
