+++ tests/unit/cmk/post_rename_site/test_load_plugins.py
@@ -7,7 +7,7 @@ from collections.abc import Sequence
 
 import pytest
 
-from tests.testlib.utils import is_cloud_repo, is_enterprise_repo
+from tests.testlib.repo import is_cloud_repo, is_enterprise_repo
 
 from cmk.post_rename_site import main
 from cmk.post_rename_site.registry import rename_action_registry
