+++ tests/unit/cmk/utils/prediction/test_services_prediction.py
@@ -9,7 +9,7 @@ import json
 
 import pytest
 
-from tests.testlib.utils import repo_path
+from tests.testlib.repo import repo_path
 
 from livestatus import RRDResponse
