+++ tests/unit/cmk/utils/structured_data/test_structured_data.py
@@ -11,7 +11,7 @@ from typing import Literal
 
 import pytest
 
-from tests.testlib.utils import repo_path
+from tests.testlib.repo import repo_path
 
 from cmk.utils.hostaddress import HostName
 from cmk.utils.structured_data import (
