+++ tests/unit/cmk/utils/test_man_pages.py
@@ -11,7 +11,7 @@ from pathlib import Path
 
 import pytest
 
-from tests.testlib.utils import repo_path
+from tests.testlib.repo import repo_path
 
 from tests.unit.conftest import FixPluginLegacy, FixRegister
